+++ pkg/sql/sem/eval/deps.go
@@ -212,8 +212,8 @@ const AdvisoryLockShared AdvisoryLockModes = 0
 const AdvisoryLockExclusive AdvisoryLockModes = 1
 
 type AdvisoryLockProvider interface {
-	AdvisoryLock(ctx context.Context, id int, mode AdvisoryLockModes, wait bool, txnScoped bool) (err error)
-	AdvisoryLockRelease(ctx context.Context, id int, mode AdvisoryLockModes) (err error)
+	AdvisoryLock(ctx context.Context, id int64, mode AdvisoryLockModes, wait bool, txnScoped bool) (err error)
+	AdvisoryLockRelease(ctx context.Context, id int64, mode AdvisoryLockModes) (err error)
 	ReleaseAllForSession(ctx context.Context) (err error)
 }
